Move Start/Stop tests to different file
